requirements.txt
# Private Compliance Agent - AML/Compliance Document & OSINT Screening
# All processing runs locally; no PII sent to external APIs.

# Orchestration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# PDF extraction (fast, local)
PyMuPDF>=1.24.0

# OSINT search (Tavily)
tavily-python>=0.7.0

# Structured output & validation
pydantic>=2.0.0

# Optional: Groq backend for fast public demos
langchain-groq>=0.2.0

# Web UI
streamlit>=1.28.0

# Env loading (.env: GROQ_API_KEY, TAVILY_API_KEY)
python-dotenv>=1.0.0
